+++ src/nationalarchives/components/cookie-banner/cookie-banner.stories.js
@@ -34,7 +34,6 @@ export default {
 const Template = ({
   serviceName,
   cookiesUrl,
-  policies,
   policiesKey,
   preferencesSetKey,
   cookiesDomain,
@@ -48,7 +47,6 @@ const Template = ({
     params: {
       serviceName,
       cookiesUrl,
-      policies,
       policiesKey,
       preferencesSetKey,
       cookiesDomain,
@@ -81,8 +79,9 @@ Accept.play = async ({ canvasElement }) => {
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("unknown")).toEqual(null);
-  await expect(cookies.exists("cookie_preferences_set")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
+  await expect(cookies.get("cookie_preferences_set")).toEqual("false");
 
   const canvas = within(canvasElement);
   const acceptButton = canvas.getByText("Accept cookies");
@@ -94,16 +93,13 @@ Accept.play = async ({ canvasElement }) => {
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("unknown")).toEqual(null);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(true);
   await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
   await expect(cookies.get("cookie_preferences_set")).toEqual("true");
-  await expect(cookies.hasValue("cookie_preferences_set", "true")).toEqual(
-    true,
-  );
   await expect(acceptButton).not.toBeVisible();
   await expect(rejectButton).not.toBeVisible();
 
-  // const closeButton = canvas.getByText("Close this message");
+  // const closeButton = canvas.getByText("Hide cookies message");
   // await expect(closeButton).toBeVisible();
   // await userEvent.click(closeButton);
 
@@ -124,8 +120,9 @@ Reject.play = async ({ canvasElement }) => {
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("unknown")).toEqual(null);
-  await expect(cookies.exists("cookie_preferences_set")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
+  await expect(cookies.get("cookie_preferences_set")).toEqual("false");
 
   const canvas = within(canvasElement);
   const acceptButton = canvas.getByText("Accept cookies");
@@ -137,7 +134,7 @@ Reject.play = async ({ canvasElement }) => {
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
   await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
   await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("unknown")).toEqual(null);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
   await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
   await expect(cookies.get("cookie_preferences_set")).toEqual("true");
   await expect(cookies.hasValue("cookie_preferences_set", "true")).toEqual(
@@ -145,45 +142,6 @@ Reject.play = async ({ canvasElement }) => {
   );
   await expect(acceptButton).not.toBeVisible();
   await expect(rejectButton).not.toBeVisible();
-
-  await cookies.deleteAll();
-};
-
-export const CustomPolicies = Template.bind({});
-CustomPolicies.args = {
-  serviceName: "My service",
-  cookiesUrl: "#",
-  policies: "custom",
-  classes: "tna-cookie-banner--demo",
-  disableMockAnalytics: true,
-};
-CustomPolicies.parameters = {
-  chromatic: { disableSnapshot: true },
-};
-CustomPolicies.play = async ({ args, canvasElement }) => {
-  const cookies = new Cookies({
-    extraPolicies: args.policies.split(","),
-    newInstance: true,
-  });
-  await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("custom")).toEqual(false);
-  await expect(cookies.exists("cookie_preferences_set")).toEqual(false);
-
-  const canvas = within(canvasElement);
-  const acceptButton = canvas.getByText("Accept cookies");
-  await userEvent.click(acceptButton);
-
-  await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("settings")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("custom")).toEqual(true);
-  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
-  await expect(cookies.get("cookie_preferences_set")).toEqual("true");
-  await expect(cookies.hasValue("cookie_preferences_set", "true")).toEqual(
-    true,
-  );
 
   await cookies.deleteAll();
 };
@@ -199,21 +157,26 @@ Existing.args = {
 Existing.decorators = [
   (Story) => {
     const cookies = new Cookies({ newInstance: true });
-    cookies.set("cookie_preferences_set", true);
+    cookies.set(
+      "cookies_policy",
+      JSON.stringify({
+        usage: true,
+        settings: true,
+        marketing: false,
+        essential: false,
+      }),
+    );
     return Story();
   },
 ];
 Existing.play = async ({ canvasElement }) => {
   const cookies = new Cookies();
   await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
-  await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
-  await expect(cookies.isPolicyAccepted("unknown")).toEqual(null);
+  await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("settings")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
   await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
   await expect(cookies.get("cookie_preferences_set")).toEqual("true");
-  await expect(cookies.hasValue("cookie_preferences_set", "true")).toEqual(
-    true,
-  );
 
   const canvas = within(canvasElement);
   const acceptButton = canvas.getByText("Accept cookies");
@@ -223,6 +186,80 @@ Existing.play = async ({ canvasElement }) => {
 
   await cookies.deleteAll();
 };
+
+export const Partial = Template.bind({});
+Partial.args = {
+  serviceName: "My service",
+  cookiesUrl: "#",
+  allowInsecure: true,
+  classes: "tna-cookie-banner--demo",
+  disableMockAnalytics: true,
+};
+Partial.decorators = [
+  (Story) => {
+    const cookies = new Cookies({ newInstance: true });
+    cookies.set(
+      "cookies_policy",
+      JSON.stringify({
+        usage: true,
+        essential: false,
+      }),
+    );
+    return Story();
+  },
+];
+Partial.play = async ({ canvasElement }) => {
+  const cookies = new Cookies();
+  await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("usage")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
+  await expect(cookies.get("cookie_preferences_set")).toEqual("false");
+
+  const canvas = within(canvasElement);
+  const acceptButton = canvas.getByText("Accept cookies");
+  const rejectButton = canvas.getByText("Reject cookies");
+  await expect(acceptButton).toBeVisible();
+  await expect(rejectButton).toBeVisible();
+
+  await cookies.deleteAll();
+};
+
+export const Malformed = Template.bind({});
+Malformed.args = {
+  serviceName: "My service",
+  cookiesUrl: "#",
+  allowInsecure: true,
+  classes: "tna-cookie-banner--demo",
+  disableMockAnalytics: true,
+};
+Malformed.decorators = [
+  (Story) => {
+    const cookies = new Cookies({ newInstance: true });
+    cookies.set("cookies_policy", "foobar");
+    return Story();
+  },
+];
+Malformed.play = async ({ canvasElement }) => {
+  const cookies = new Cookies();
+  await expect(cookies.isPolicyAccepted("essential")).toEqual(true);
+  await expect(cookies.isPolicyAccepted("usage")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("settings")).toEqual(false);
+  await expect(cookies.isPolicyAccepted("marketing")).toEqual(false);
+  await expect(cookies.exists("cookie_preferences_set")).toEqual(true);
+  await expect(cookies.get("cookie_preferences_set")).toEqual("false");
+
+  const canvas = within(canvasElement);
+  const acceptButton = canvas.getByText("Accept cookies");
+  const rejectButton = canvas.getByText("Reject cookies");
+  await expect(acceptButton).toBeVisible();
+  await expect(rejectButton).toBeVisible();
+
+  await cookies.deleteAll();
+};
+
+//Hidden w. no pref saved
 
 // export const EventHandling = Template.bind({});
 // EventHandling.args = {